Pull some common DI boiler plate out into an abstract class.
